requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
docopt==0.6.2
gitignore==0.0.8
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2018.9
static3==0.7.0
whitenoise==4.1.2
djangorestframework
pillow
